Add backend Flask files
To be commited is a folder containing the backend of this project,
including some very basic files for running Flask. Virtual environments are never commited
to GitHub for good reason (these should be generated on the developer's
computer), and instructions for setting one up for this project in the Wiki under Wiki ->
Project Configuration.

./backend/requirements.txt streamlines setting up your own virtual
environment with all of the dependencies needed to run the code commited
here.

Currently, index.html with app.py just displays a submission box for
passing data to Flask. This is basically the least I could get away with
testing other functionality with, i.e. passing URLs to Flask for
processing.