Add pseudocache functionality
The pseudocache is a folder containing HTML .txt files corresponding to
previous URLs we have made requests to. Each file's name is a SHA-256
hash of its URL to ensure that (1) they are composed on allowed
characters (2) each file is given a unique name.

The primary purpose of the pseudocache is reducing how many requests we
send to test URLs. Notably, it does not elimiate the necessity of GET
requests since many assets are purposefully designed to load late by
many web developers.

The secondary purpose is to use this as a "proof of concept," because
loading these pages from memory is noticeably faster than querying for
them every time we submit a URL.