Convert app.py into an API interface
Previously, I was testing Flask as a standalone program using templates.
I have modified its routing so that it is an API and can get requests.
Future testing should be done using Postman, Insomnia, or some other
software for testing APIs (I am using Insomnia at present).

This allows me to add a frontend linked to the WASA Flask API such that
both components are as independent as possible.