Provide barebones Angular/Flask app
This required modification of backend and frontend to work.

On backend, I had to enable CORS to return data to the frontend.

On the frontend, I provide two service classes. HandleUrlService is for
URL validation. ConfigService configures HTTP requests to the Flask API.

Additionally, on the frontend we now have the following components: app,
landing, url-submission-form, and dynamic-article. App routes to
landing, and landing just prints url-submission-form and
dynamic-article. Url-submission-form taks a URL, validates it, and makes
a request to the Flask API if the URL is valid (logging the response in
the console). Finally, dynamic-article is just static right now (it
needs to be expanded to use the data from the Flask API to build an
article.)