Add styling
Frontend is styled with Bootstrip 5 elements, where the font has been
overridden to use an accessible font provided freely by the Braille
Institute are brailleinstititute.org.

Care has been taken to make the application easy to read (for low vision
users), possible to navigate with a screen reader, and easy to use, but
these are ongoing efforts; one problem that exists now? Contrast in the
main URL submission form button is too low.