Flatten scrape_url()
Main method of the Flask API was awful to read through. Truly a hackjob
of nested code. I removed these nests ("flattened") it to essentially
have a "check section" and "business logic" section.

The purpose of doing this is to make dealing with the rest of the logic
easier. It would be nice to be able to have separate modes for parsing
HTML using BS4, Trafilatura, etc., but that needs refactoring which
would be too hard with the confusing scrape_url() method.