Add support for Trafilatura article parsing
Trafilatura is a purpose-built text extraction Python library. WASA can
operate in two modes. Mode 0 uses just BS4 parsing (our original
approach). Mode 1 uses Trafilatura to capture text, and BS4 to simply format
what we grab.

Without testing this on a variety of articles, there is no way to say
whether this is a "better" approach than just using BS4. On the limited
number of test articles I ran through WASA, it looks like we capture
less noise. A better approach would perhaps be to write a test script or
to repurpose the Trafilatura authors' test script to a set of my own
articles.

Speed does not seem to be a concern, but once a few modes have been
incorporated into the project we might do some benchmarking tests on
which parser is the fastest on some set of articles.