Add themes
There are four themes: seaside (lightmode), astral (darkmode), pickle
(earthy), and apricot (vibrant). At first I wanted to do this with Tailwinds, but Tailwinds only natively supports switching between lightmode and darkmode. There are plugins to use more than 2 fonts, but I did not want to use a plugin.

Instead, we use Angular ngClass in the landing component to set a custom
CSS class. This class is set on each component we apply our themes to.

Each theme has a primary, secondary, accent, and highlight colour. The CSS rules for these themes are in styles.css.

The colours used in these themes at present should be thought of as a
starting point. They need to be tested for colourblindness
accessibility. They also need to be refined aesthetically AND for
contrast between the colours themselves, not just the colours and text.
For example, highlight should have high contrast with both its text
colour AND the secondary colour for each theme.

There are also some obvious improvements to be made on the buttons. In
particular, each button does not change to show the user which theme
they have selected. They also don't hint at what the theme looks like,
e.g. by being styled as the theme they are.