Add wasa_db_handlr.py, a database handling module for WASA
This is a major update adding a new module to the project for brokering
data between the main module, app.py, and a Postgres database.

There is still a lot to do with this module. It needs to be reviewed,
refactored, and integrated into app.py to replace the current
"pseudocache." This also means that most methods in the module need to
be set to private functions.

However, this work is to be done later because this commit includes some
important working test code for performing writes, shallow deletes, deep
deletes, and reading from the cache. It also repeats some code from
app.py for the purposes of testing. These parts need removed from the
module.

The intended behaviour of the module is to handle all of the logic
required for reading and writing to the cache, which is a Postgres SQL
database. There are two intended access points: a function for reading,
and a function for writing. Other rules, such as staleness rules, would
be included in this module (NOT app.py). This module might also be
expanded to allow an admin to query the database for some data, such as
request records.

The database caches JSON. It needs to be updated to collect request
records (this is a simple change, but it is not pressing.)