Format wasa_db_handlr.py
Removed tests. Also moved any code that would execute when running the
module to the top (these will be moved elsewhere since we can't use the
module itself to open and use the database connection.

Added _ as a prefix for "private" functions. Python doesn't REALLY have
private functions, but this indicates to the programmer that they
shouldn't call these functions.

This has been said elsewhere: but the whole idea of the handler is that
it should be called from its read/write functions.