Integrate wasa_db_handler.py with app.py
We have officially moved on from using the "pseudocache!" app.py is
configured to use wasa_db_handler.py whenever it wants to read or write
from the cache.

Requests are noticeably faster. I think that read/writing to the file
system was slow, perhaps in-part because my jury-rigged solution was to
save each file with a SHA256 hash.

Another reason might be that before, files were saved as HTML and parsed
on every response. Requests for cached URLs are ALSO faster, so since
parsing is an expensive operation we could be saving time be caching
just the PARSED webpage.

Now, the system tries to read from the Postgres DB. If it exists, this
is sent to the user. If it does not, only then do we make a request to
the URL, parse it, cache it, and return the JSON to the user.