Add functionality for reading .txt into dictionaries
I am working on adding image scraping to the API. The problem with image
scraping is that doing it poorly is easy, and doing it well by omitting
irrelevant images is hard!

The approach I am playing around with is using heuristics to exclude
irrelevant images while alos identifying the most import images. One of
these heuristics is whether an image is pulled from an adserver.
Searching a list every time I do this would be inefficient, so we need
functionality to save that list into main memory using a
dictionary. Finding a domain name in a txt file is O(n), whereas finding
it in a dictionary is O(1). Yes, it is O(n) to build the dictionary, but
we might need to search it 1-(# images) times per query, so it's still faster.

The code in domain_parser.py tries to read the adserver dictionary from a pickled file.
If it doesn't exist, it builds the dictionary from a txt file and saves
the dictionary object as a pickle. Finally, to prove this is faster than rebuilding the dictionary every request, I have added a test function in function_tests.py comparing the runtime of the two.

I wanted to add functionality for an allowlist and blocklist anyway down
the line, so it makes sense to add functionality to build these dictionaries
now. These have the same functional requirements as reading from the aservers
list (i.e. that we will do it frequently ... so we need to do it quickly.)