Add support for images
The good news is that WASA supports images now! The bad news is that
there's a lot to improve...

I had to edit the frontend and backend (app.py AND wasa_db_handler) to
get everything working. There were some logic errors in wasa_db_handler
that made webpage elements load out of order.

I think that the next step would be improving on how text and images are
processed, or adding a "linter" with some hardcoded linting logic on the
text and images. Sort of like stopword removal for text, or using some
common heuristics on images (e.g. "only keep the largest image").