Replace <div> elements with better semantic HTML
<div> should be avoided wherever possible for elements that describe
these sections better. Almost all of my <div> elements have been changed
to <article>, <nav>, or <form>. However: this is a breaking change on
the URL submission form. I need to handle the data there in a different
way such that the data is pulled instead of submitted through the form.
(Or rework how the applications submits that data!)