Move options to submission form
I am in the process of properly submitting the URL using a form, and it
makes more sense for the user options to be a part of this form. The
"options" card is now just a theme card, but it could be expanded to
support other customization options.

i.e. this is a more logical grouping for the data.