Add support for alt-text
The WAPI badly needs some refactoring. I had to hunt through a few
different functions meant to produce JSON before finding the culprit
that prevented the frontend from properly digesting alt-text.

Before that, I need to take a look at fixing how data is saved to the
databse. Images in particular save alt-text to the 'caption' column when
it should be saved to the 'alt_text' column!