Add get_images, generate_alt_text optional arguments
Callers can now specify whether to get images or not, and whether to get
generated alt-text or not from the API. These are treated as True,
False, by default if they are not specified. Main wiki has more
information in section C.

The API doesn't respect these flags yet, it's just capable of taking
them as arguments.