Add foundational support for get_images, generate_alt_text
json_linter has been modified to account for get_images and
generate_alt_text when linting some generated json.

I intend to combine these functions because work is unnecessarily being
done twice. Both functions are very similar, but this will require some
care. It might also mean taking a look at how data is saved to the
database, since I am not 100% sure if the database handler can handle
generated alt-text or not.

Anothing thing that needs done is changing when the API writes to the
database. Right now, this write is done before returning data to the
caller. This should instead be handled by @app.after_request so that the
user isn't waiting on data being written to the database. An added
benefit: if there is a write failure, the user will still get the JSON
they wanted.