Consolidate parse_response, write_json, json_linter -> parse_response
All of these functions existed for the same essential purpose: parsing
the BS5/Trafilatura collection into the JSON response. It makes sense to
finally group them.

Here's the idea: use parse_response when we are scraping a webpage, and
use wasa_db_handler to read this in the future. Place code for
generating alt-text in generate_alt_text_form_bs4; so long as it return
a string, this will not break the rest of the app. wasa_db_handler is
disabled right now; the next problem is to